Add a custom sphinx extension
The custom extension comes with a 'overview' admonition that can
be used for an overview at the start of every section, and a 'findoutmore'
directive that creates foldable section. Both are taken from the
DataLad handbook.

The custom sphinx extension requires some setup to allow installation.
This is done with a modern setup.py using versioneer.

.gitignore is adjusted to ignore created python files created during
the build

Typo in gitignore file